refactor(user): DTO 및 서비스 역할별 분리
- dto/user.dto.ts — 일반 유저용 DTO (CreateUserDto, SubmitRegistrationDto, UpdateMyInfoDto)
- dto/user-admin.dto.ts — 관리자용 DTO (UpdateUserInfoDto, UserListFilter, FindFilteredResult)
- service/user-admin.service.ts — 관리자 전용 서비스 (승인·거절·목록·편집·비활성화)
- service/user-class-rep.service.ts — 반대표 전용 서비스 (자기 반 승인 대기·목록 조회)
- user.service.ts — 공통 메서드만 유지 (조회·가입 흐름·토큰·내 정보 수정)
